Revert "Farid"
