+++ Assets/Scenes/Main Menu.unity
@@ -2251,64 +2251,3 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1844399982}
   m_CullTransparentMesh: 1
---- !u!1001 &936092975149731941
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: 936092974423396882, guid: eb0c3102f4f8564478bfd38146513212, type: 3}
-      propertyPath: m_Name
-      value: SongSpawner
-      objectReference: {fileID: 0}
-    - target: {fileID: 936092974423396882, guid: eb0c3102f4f8564478bfd38146513212, type: 3}
-      propertyPath: m_IsActive
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 936092974423396892, guid: eb0c3102f4f8564478bfd38146513212, type: 3}
-      propertyPath: m_RootOrder
-      value: 5
-      objectReference: {fileID: 0}
-    - target: {fileID: 936092974423396892, guid: eb0c3102f4f8564478bfd38146513212, type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 412
-      objectReference: {fileID: 0}
-    - target: {fileID: 936092974423396892, guid: eb0c3102f4f8564478bfd38146513212, type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 1475
-      objectReference: {fileID: 0}
-    - target: {fileID: 936092974423396892, guid: eb0c3102f4f8564478bfd38146513212, type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 936092974423396892, guid: eb0c3102f4f8564478bfd38146513212, type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 936092974423396892, guid: eb0c3102f4f8564478bfd38146513212, type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 936092974423396892, guid: eb0c3102f4f8564478bfd38146513212, type: 3}
-      propertyPath: m_LocalRotation.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 936092974423396892, guid: eb0c3102f4f8564478bfd38146513212, type: 3}
-      propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 936092974423396892, guid: eb0c3102f4f8564478bfd38146513212, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 936092974423396892, guid: eb0c3102f4f8564478bfd38146513212, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 936092974423396892, guid: eb0c3102f4f8564478bfd38146513212, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: eb0c3102f4f8564478bfd38146513212, type: 3}
