--- Assets/SongObject/SongSpawner.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: eb0c3102f4f8564478bfd38146513212
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
