--- Assets/Tables.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 32e8bd3b5e248554388a45dfa5c076e9
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
