--- Assets/Tables/Table 1.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 3bfe2cec08922424d91f2307ae1c2ab3
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
