--- Assets/Tables/Table.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6b4ad9a743f778c4886ee6edf2ddadc4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
