--- Assets/Tables/UIFollowObject.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fc4a8d984fb87e1438e020b381a0ce8b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
