+++ ProjectSettings/TagManager.asset
@@ -6,7 +6,7 @@ TagManager:
   tags:
   - Player1
   - Player2
-  - ReadyText
+  - SongBar
   layers:
   - Default
   - TransparentFX
