+++ UserSettings/EditorUserSettings.asset
@@ -30,10 +30,10 @@ EditorUserSettings:
       value: 55570d0007060a0c5b5d5f7142710b444f15192b7a7e74657a7e1b31bae23561
       flags: 0
     RecentlyUsedSceneGuid-8:
-      value: 5a050255065608035a0b0a2712710d44144e482c742e7f652e7c4831b2e5316d
+      value: 005507045c570b0b595a5d7214740c4447154f2b2d2b77332c701866b7b7376c
       flags: 0
     RecentlyUsedSceneGuid-9:
-      value: 005507045c570b0b595a5d7214740c4447154f2b2d2b77332c701866b7b7376c
+      value: 5a050255065608035a0b0a2712710d44144e482c742e7f652e7c4831b2e5316d
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
